+++ src/utils/user.ts
@@ -1,0 +1,8 @@
+export type User = {
+  userTelegramID: string;
+  userName: string;
+  userHandle: string;
+  patchwallet: string;
+  dateAdded: string;
+  photoUrl?: string;
+};